Revert "Fix ByteUtils import after Kafka client relocation (#4232)"
This reverts commit 306a80a4413ba8a93be1d1b056e6852b54c9870d.
