+++ protobuf-provider/src/main/java/io/confluent/kafka/schemaregistry/protobuf/MessageIndexes.java
@@ -15,7 +15,7 @@
 
 package io.confluent.kafka.schemaregistry.protobuf;
 
-import org.apache.kafka.common.utils.internals.ByteUtils;
+import org.apache.kafka.common.utils.ByteUtils;
 
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
